implemented filters for email scraping inside the same view
